backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
django-cors-headers==4.3.0
drf-yasg==1.21.7 